[Rust Frontend] Extract shared SslArgs from frontend and render args
Both SharedRuntimeArgs and RenderArgs carried identical ssl_* field
definitions and tls_config() builders. Move them into a shared
cli::ssl::SslArgs struct flattened into both, keeping the CLI flags
and the Python-supervisor JSON contract unchanged.
