+++ rust/src/cmd/src/cli.rs
@@ -7,6 +7,7 @@
 //! - Engine args: <https://github.com/vllm-project/vllm/blob/bc2c0c86efb28e77677a3cfb8687e976914a313a/vllm/engine/arg_utils.py#L657-L1311>
 //! - Environment variables: <https://github.com/vllm-project/vllm/blob/bc2c0c86efb28e77677a3cfb8687e976914a313a/vllm/envs.py#L472>
 
+mod ssl;
 mod unsupported;
 
 use std::collections::HashMap;
@@ -30,9 +31,9 @@ use vllm_managed_engine::cli::{ManagedEngineArgs, repartition_managed_engine_arg
 use vllm_server::{
     ApiServerOptions, ChatTemplateContentFormatOption, Config, CoordinatorMode, CorsConfig,
     DEFAULT_KEEP_ALIVE_TIMEOUT, HttpListenerMode, ParserSelection, RenderConfig, RendererSelection,
-    TlsConfig,
 };
 
+use crate::cli::ssl::SslArgs;
 use crate::cli::unsupported::UnsupportedArgs;
 
 /// Top-level parser for the `vllm-rs` binary.
@@ -143,44 +144,14 @@ pub struct RenderArgs {
     /// Maximum accepted logprobs count; -1 disables the cap.
     #[arg(long, value_parser = clap::value_parser!(i32).range(-1..), allow_negative_numbers = true)]
     max_logprobs: Option<i32>,
-    /// The file path to the SSL key file. When omitted, the key is read from `--ssl-certfile` (combined PEM).
-    #[arg(long)]
-    ssl_keyfile: Option<String>,
-    /// The file path to the SSL cert file. Enables TLS when set.
-    #[arg(long)]
-    ssl_certfile: Option<String>,
-    /// The CA certificates file used to verify client certificates (mTLS).
-    #[arg(long)]
-    ssl_ca_certs: Option<String>,
-    /// Whether a client certificate is required: 0 = none, 1 = optional,
-    /// 2 = required (mirrors Python's `ssl.CERT_*`).
-    #[arg(long, default_value_t = 0, value_parser = clap::value_parser!(i32).range(0..=2))]
-    ssl_cert_reqs: i32,
-    /// OpenSSL cipher string for HTTPS (TLS 1.2 and below).
-    #[arg(long)]
-    ssl_ciphers: Option<String>,
+    /// TLS options for HTTPS/mTLS.
+    #[command(flatten)]
+    ssl: SslArgs,
 }
 
 impl RenderArgs {
-    /// Build the TLS config: `Some` when any `ssl_*` argument is set, else `None` (plaintext).
-    // The combination is validated in [`RenderConfig::validate`].
-    fn tls_config(&mut self) -> Option<TlsConfig> {
-        let tls_requested = self.ssl_certfile.is_some()
-            || self.ssl_keyfile.is_some()
-            || self.ssl_ca_certs.is_some()
-            || self.ssl_cert_reqs != 0
-            || self.ssl_ciphers.is_some();
-        tls_requested.then(|| TlsConfig {
-            cert_file: self.ssl_certfile.take(),
-            key_file: self.ssl_keyfile.take(),
-            ca_certs: self.ssl_ca_certs.take(),
-            cert_reqs: self.ssl_cert_reqs,
-            ciphers: self.ssl_ciphers.take(),
-        })
-    }
-
-    pub(super) fn into_config(mut self) -> RenderConfig {
-        let tls = self.tls_config();
+    pub(super) fn into_config(self) -> RenderConfig {
+        let tls = self.ssl.tls_config();
 
         RenderConfig {
             model: self.model,
@@ -389,33 +360,10 @@ pub struct SharedRuntimeArgs {
     #[serde(default)]
     pub allow_credentials: bool,
 
-    /// The file path to the SSL key file. When omitted, the key is read from
-    /// `--ssl-certfile` (combined PEM).
-    #[arg(long)]
-    #[serde(default)]
-    pub ssl_keyfile: Option<String>,
-
-    /// The file path to the SSL cert file. Enables TLS when set.
-    #[arg(long)]
-    #[serde(default)]
-    pub ssl_certfile: Option<String>,
-
-    /// The CA certificates file used to verify client certificates (mTLS).
-    #[arg(long)]
-    #[serde(default)]
-    pub ssl_ca_certs: Option<String>,
-
-    /// Whether a client certificate is required: 0 = none, 1 = optional,
-    /// 2 = required (mirrors Python's `ssl.CERT_*`).
-    #[arg(long, default_value_t = 0, value_parser = clap::value_parser!(i32).range(0..=2))]
-    #[serde(default)]
-    pub ssl_cert_reqs: i32,
-
-    /// OpenSSL cipher string for HTTPS (TLS 1.2 and below).
-    /// When unset, the linked OpenSSL's default suites are used.
-    #[arg(long)]
-    #[serde(default)]
-    pub ssl_ciphers: Option<String>,
+    /// TLS options for HTTPS/mTLS.
+    #[command(flatten)]
+    #[serde(default, flatten)]
+    pub ssl: SslArgs,
 
     /// Profiler configuration forwarded by the Python supervisor.
     ///
@@ -507,7 +455,7 @@ impl SharedRuntimeArgs {
         let keep_alive_timeout = self.keep_alive_timeout();
         let api_server_options = self.api_server_options();
         let cors = self.cors_config();
-        let tls = self.tls_config();
+        let tls = self.ssl.tls_config();
         let profiler = self.profiler();
 
         Config {
@@ -564,7 +512,7 @@ impl SharedRuntimeArgs {
         let keep_alive_timeout = self.keep_alive_timeout();
         let api_server_options = self.api_server_options();
         let cors = self.cors_config();
-        let tls = self.tls_config();
+        let tls = self.ssl.tls_config();
         let profiler = self.profiler();
 
         Config {
@@ -617,23 +565,6 @@ impl SharedRuntimeArgs {
             allow_headers: self.allowed_headers.0.clone(),
             allow_credentials: self.allow_credentials,
         }
-    }
-
-    /// Build the TLS config: `Some` when any `ssl_*` argument is set, else
-    /// `None` (plaintext). The combination is validated in [`Config::validate`].
-    fn tls_config(&self) -> Option<TlsConfig> {
-        let tls_requested = self.ssl_certfile.is_some()
-            || self.ssl_keyfile.is_some()
-            || self.ssl_ca_certs.is_some()
-            || self.ssl_cert_reqs != 0
-            || self.ssl_ciphers.is_some();
-        tls_requested.then(|| TlsConfig {
-            cert_file: self.ssl_certfile.clone(),
-            key_file: self.ssl_keyfile.clone(),
-            ca_certs: self.ssl_ca_certs.clone(),
-            cert_reqs: self.ssl_cert_reqs,
-            ciphers: self.ssl_ciphers.clone(),
-        })
     }
 }
