+++ rust/src/cmd/src/cli/tests.rs
@@ -195,11 +195,13 @@ fn serve_args_forward_python_flags_with_separator() {
                             ],
                         ),
                         allow_credentials: false,
-                        ssl_keyfile: None,
-                        ssl_certfile: None,
-                        ssl_ca_certs: None,
-                        ssl_cert_reqs: 0,
-                        ssl_ciphers: None,
+                        ssl: SslArgs {
+                            ssl_keyfile: None,
+                            ssl_certfile: None,
+                            ssl_ca_certs: None,
+                            ssl_cert_reqs: 0,
+                            ssl_ciphers: None,
+                        },
                         profiler_config: None,
                     },
                     managed_engine: ManagedEngineArgs {
@@ -914,11 +916,13 @@ fn frontend_args_accept_json() {
                             ],
                         ),
                         allow_credentials: false,
-                        ssl_keyfile: None,
-                        ssl_certfile: None,
-                        ssl_ca_certs: None,
-                        ssl_cert_reqs: 0,
-                        ssl_ciphers: None,
+                        ssl: SslArgs {
+                            ssl_keyfile: None,
+                            ssl_certfile: None,
+                            ssl_ca_certs: None,
+                            ssl_cert_reqs: 0,
+                            ssl_ciphers: None,
+                        },
                         profiler_config: None,
                     },
                 },
@@ -1508,11 +1512,13 @@ fn serve_args_accept_handshake_aliases() {
                             ],
                         ),
                         allow_credentials: false,
-                        ssl_keyfile: None,
-                        ssl_certfile: None,
-                        ssl_ca_certs: None,
-                        ssl_cert_reqs: 0,
-                        ssl_ciphers: None,
+                        ssl: SslArgs {
+                            ssl_keyfile: None,
+                            ssl_certfile: None,
+                            ssl_ca_certs: None,
+                            ssl_cert_reqs: 0,
+                            ssl_ciphers: None,
+                        },
                         profiler_config: None,
                     },
                     managed_engine: ManagedEngineArgs {
